feat: search history, watchlist tab, and fix comps NaN
NaN fix:
- ComparableSale schema: Decimal fields (price, baths, price_per_sqft)
  changed to float — Pydantic v2 serializes Decimal as strings, breaking
  the JS reduce() that computes avg price in CompsPanel

Search history:
- SearchBar reads/writes recent queries to localStorage under
  landgrab:search_history (max 8 entries)
- Dropdown appears on focus showing matching history items with clock icon
- onMouseDown handler prevents blur from closing before navigation fires

Watchlist (favorites) tab:
- /favorites page: grid of saved properties with price-change indicator
  (current vs snapshot), remove-on-hover trash button, beds/baths/sqft
- Watchlist nav link added to Navbar after Map
- Save button in PropertyDetailClient now calls api.saveProperty() via
  useMutation; fills heart icon and shows "Saved" state on success
- user_service.py: all flush() calls replaced with commit() so saves,
  deletes, and search writes actually persist
- list_saved_properties now returns zip_code, beds, baths, sqft fields

+++ backend/app/schemas/property.py
@@ -29,13 +29,13 @@ class ComparableSale(BaseModel):
     address: str
     city: str
     state: str
-    price: Decimal
+    price: float
     sqft: Optional[int] = None
     beds: Optional[int] = None
-    baths: Optional[Decimal] = None
+    baths: Optional[float] = None
     sale_date: date
     distance_miles: Optional[float] = None
-    price_per_sqft: Optional[Decimal] = None
+    price_per_sqft: Optional[float] = None
     similarity_score: Optional[float] = None
